Moving logic to middleware
