+++ src/middleware.ts
@@ -1,5 +1,6 @@
-import { sort } from './lib/utils';
+import config from './config';
 import mapper from './lib/mapper';
+import { sort } from './lib/utils';
 import BurritoStore from './store/BurritoStore';
 
 /**
@@ -11,9 +12,60 @@ import BurritoStore from './store/BurritoStore';
  * @param {string} listType - to / from
  */
 const getScoreBoard = async (scoreType: string, listType: string) => {
-    const scoreList = await BurritoStore.getScoreBoard({ listType, scoreType });
+    const data = await BurritoStore.getScoreBoard({ listType, scoreType });
+
+    const score = [];
+    const uniqueUsername = [...new Set(data.map((x) => x[listType]))];
+
+    const { enable_decrement } = config.slack
+
+    const scoreTypeFilter = (scoreType === 'inc') ? 1 : -1;
+
+    uniqueUsername.forEach((u) => {
+        const dataByUser = data.filter((e: any) => (e[listType] === u));
+        let filteredData: any;
+        let countSwitch: any;
+
+        if (listType === 'to' && enable_decrement && (scoreType === 'inc')) {
+            filteredData = dataByUser
+
+        } else {
+            filteredData = dataByUser.filter((e: any) => (e.value === scoreTypeFilter));
+            countSwitch = 1
+        }
+        const red = filteredData.reduce((a: number, item) => {
+            return a + (countSwitch || item.value);
+        }, 0);
+        score.push({ _id: u, score: red });
+    });
+    const scoreList = score.map(x => {
+        if (x.score != 0) return x
+        return undefined
+    }).filter(y => y)
+
     return sort(mapper(scoreList));
 };
+
+
+const getUserScoreBoard = async ({ ...args }) => {
+    const { listType } = args;
+    const data: any = await BurritoStore.getScoreBoard({ ...args });
+    const score = [];
+    const uniqueUsername = [...new Set(data.map((x) => x[listType]))];
+    uniqueUsername.forEach((u) => {
+        const dataByUser = data.filter((e: any) => e[listType] === u);
+        const scoreinc = dataByUser.filter((x: any) => x.value === 1);
+        const scoredec = dataByUser.filter((x: any) => x.value === -1);
+        score.push({
+            _id: u,
+            scoreinc: scoreinc.length,
+            scoredec: scoredec.length,
+        });
+
+    });
+    return score;
+}
+
 
 /**
  * @param {string} user - Slack userId
@@ -27,10 +79,10 @@ const getUserStats = async (user: string) => {
         receivedListToday,
     ] = await Promise.all([
         BurritoStore.getUserStats(user),
-        BurritoStore.getUserScoreBoard({ user, listType: 'to' }),
-        BurritoStore.getUserScoreBoard({ user, listType: 'from' }),
-        BurritoStore.getUserScoreBoard({ user, listType: 'to', today: true }),
-        BurritoStore.getUserScoreBoard({ user, listType: 'from', today: true }),
+        getUserScoreBoard({ user, listType: 'to' }),
+        getUserScoreBoard({ user, listType: 'from' }),
+        getUserScoreBoard({ user, listType: 'to', today: true }),
+        getUserScoreBoard({ user, listType: 'from', today: true }),
     ]);
 
     return {
